feat: New boss solving algorithm
Solver works in two phases.

In phase 1 we find all potential solutions (i.e. combinations of traits)
 for each node and crew that matches those solutions. At the end of
phase 1 we eliminate all solutions that have only one matching crew (as
this is an official rule) and also all solutions that are covered by
already tried crew.

In phase 2 we try to improve solutions by finding traits that must be
included in the correct solution for each node and eliminating now
impossible solutions for other nodes. This runs in a loop until no new
solved traits are found.

The solver doesn't use any of the unofficial rules, but it is easy to
add them.